fix: convert all normalize stage patchers to TypeScript (#1134)
Some of these could be done independently, but there were about 7 files that had
references between each other. Generally these were straightforward, although I
reworked the control flow a little in a few places, and I had to make
NodePatcher aware of the "tack a callback on so a child can use it" approach
that FunctionPatcher does.

Mark as fix to force a release so we don't go too long without running the full
test suite of projects.

This also involved changing the typescript settings to allow doing for-of loops
and spreads over iterables using the downlevelIteration option. This introduced
usages of new TypeScript helpers that caused a crash when running the binary. I
think the rollup typescript plugin was trying to add all helpers, but it had an
out-of-date list, so instead I configured TypeScript to import from tslib
instead, then I added it as a dependency.