fix: properly downgrade unicode code point escapes in regexes (#1237)
This fixes the last CoffeeScript test failure after upgrading to 1.12.7.

CoffeeScript specifically handles unicode code point escapes in regexes without
the "u" flag specified, e.g. `/\u{a}/`, whereas JavaScript requires the "u" flag
for it to be interpreted as a code point escape. Unfortunately, it's incorrect
to just add the "u" flag, since that introduces various other semantics. To fix,
we just copy CoffeeScript's behavior and downgrade to old-style unicode escapes
in the BMP case and emit old-style unicode escapes for the surrogate pair in the
non-BMP case. When the "u" flag is specified, we skip this behavior.